requirements.txt
appnope==0.1.2
autopep8==1.5.4
backcall==0.2.0
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
chroma-py==0.1.0.dev1
colorama==0.4.3
colour==0.1.5
cryptography==3.4.6
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
docutils==0.16
geojson==2.5.0
gitdb==4.0.7
GitPython==3.1.18
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.2
jira==2.0.0
jupyter-client==6.1.12
jupyter-core==4.7.1
keyring==21.3.0
kiwisolver==1.2.0
mapboxgl==0.10.2
MarkupSafe==1.1.1
matplotlib==3.3.0
matplotlib-inline==0.1.2
numpy==1.19.1
oauthlib==3.1.0
packaging==20.4
parso==0.8.2
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkginfo==1.5.0.1
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.6.1
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
pyzmq==22.1.0
readme-renderer==26.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
six==1.15.0
smmap==4.0.0
toml==0.10.1
tornado==6.1
tqdm==4.48.2
traitlets==5.0.5
twine==3.2.0
typing-extensions==3.10.0.0
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.1.0
